Historia venta entradas casi finalizada.

+++ IPS2024-PL53/src/main/java/backend/service/ventas/Venta.java
@@ -1,10 +1,12 @@
 package backend.service.ventas;
 
-import java.sql.Date;
+
+import java.util.Date;
 
 public interface Venta {
 	
-	public double getPrecio();
+	public float getPrecio();
+	
 	public Date getFecha();
 
 }
